requirements.txt
textwrap
numpy
pandas
google.generativeai
streamlit
unicodedata
pymupdf
PyPDF2